Add extra padding when the status bar isn't hidden. (#4)
With the buttons having a background now, they were squished up next to the iOS status bar and looked kind of awkward. Ideally we'd hide the status bar while the lightbox is open, but that doesn't work in Shortwave because React Native is controlling the status bar. So I'm adding some padding to make them look better.